requirements.txt
asgiref==3.5.2
asttokens==2.0.7
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
black==22.6.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
executing==0.10.0
gunicorn==20.1.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jsonschema==4.16.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.12.0
PyJWT==2.5.0
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
tzdata==2022.5
uritemplate==4.1.1
waitress==2.1.2
wcwidth==0.2.5
